requirements.txt
# Core dependencies
langchain>=0.1.4
langchain-openai>=0.0.4
langchain-community>=0.0.12
langchain-core>=0.1.14
openai>=1.3.0

# Utility libraries
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.5.0

# Vector databases
chromadb>=0.4.18

# Tools dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
tavily-python>=0.2.6   # Optional, for web search

# Optional components
huggingface-hub>=0.19.4  # For alternative embeddings
sentence-transformers>=2.2.2  # For alternative embeddings

# Development and testing
pytest>=7.4.3
black>=23.11.0
